Working client.
First, a scheme needs to be created and populated with both k8s native
types and then with custom types (e.g. openshift apis).

The client is an interface, and there's a separate one with Watch
method.

It's worth mentioning that types can be added (and used wigh Get/List)
to the scheme after the client was created.